fix(html-preview): extract HTML preview frame handlers
